Add files via upload
Screenshots of the Stock Ticker results on the multiyear Stock Data file.  The VBA runs through all three sheets in one Macro, showing the Yearly Difference - difference between the beginning of the year and the end of the year - and the Percentage Difference between the beginning of the year and the end of the year.  If the Yearly Difference is positive, it is highlighted in Green, Negative Difference is highlighted in Red.